fcn8s config

+++ ptsemseg/loss/loss.py
@@ -49,7 +49,7 @@ def bootstrapped_cross_entropy2d(input, target, K, weight=None, size_average=Tru
 
         n, c, h, w = input.size()
         input = input.transpose(1, 2).transpose(2, 3).contiguous().view(-1, c)
-        target = target.view(-1)
+        target = target.view(-1).long()
         loss = F.cross_entropy(
             input, target, weight=weight, reduce=False, size_average=False, ignore_index=250
         )
